rt, refactor: extract basic_scheduler::Config
The `basic_scheduler` uses a `Config` struct to pass in runtime
configuration options. These config options are set at the public
`runtime::Builder` level and are shared between the basic scheduler and
the multi-threaded scheduler.

This change extracts the struct and uses it to configure the
multi-threaded scheduler as well
