+++ tokio/src/runtime/basic_scheduler.rs
@@ -5,7 +5,7 @@ use crate::park::{Park, Unpark};
 use crate::runtime::context::EnterGuard;
 use crate::runtime::driver::Driver;
 use crate::runtime::task::{self, JoinHandle, OwnedTasks, Schedule, Task};
-use crate::runtime::{Callback, HandleInner};
+use crate::runtime::{Config, HandleInner};
 use crate::runtime::{MetricsBatch, SchedulerMetrics, WorkerMetrics};
 use crate::sync::notify::Notify;
 use crate::util::atomic_cell::AtomicCell;
@@ -66,24 +66,6 @@ struct Core {
 #[derive(Clone)]
 pub(crate) struct Spawner {
     shared: Arc<Shared>,
-}
-
-pub(crate) struct Config {
-    /// How many ticks before pulling a task from the global/remote queue?
-    pub(crate) global_queue_interval: u32,
-
-    /// How many ticks before yielding to the driver for timer and I/O events?
-    pub(crate) event_interval: u32,
-
-    /// Callback for a worker parking itself
-    pub(crate) before_park: Option<Callback>,
-
-    /// Callback for a worker unparking itself
-    pub(crate) after_unpark: Option<Callback>,
-
-    #[cfg(tokio_unstable)]
-    /// How to respond to unhandled task panics.
-    pub(crate) unhandled_panic: crate::runtime::UnhandledPanic,
 }
 
 /// Scheduler state shared between threads.
